+++ tests/test_parser.py
@@ -3,8 +3,7 @@ from copy import deepcopy
 from pathlib import Path
 
 import pytest
-from dictIO.cppDict import CppDict
-from dictIO.parser import CppParser, Parser, XmlParser
+from dictIO import CppDict, CppParser, Parser, XmlParser
 from dictIO.utils.counter import BorgCounter
 from dictIO.utils.strings import string_diff